Revert "tried a bunch of optimizations"
This reverts commit a9c2f2c07908c87e77e70fabc88f0d896ff9a017.

+++ src/main/scala/BildScript/Generators.scala
@@ -4,14 +4,14 @@ import scala.language.implicitConversions
 
 object Generators {
 
-  case class CountGen(currentNumber: Float, stepFunc: Float => Float) extends Gen[Float] {
-    override def nextGen: Gen[Float] = CountGen(stepFunc(currentNumber), stepFunc)
-    override def get: Float = currentNumber
+  case class CountGen(currentNumber: Double, stepFunc: Double => Double) extends Gen[Double] {
+    override def nextGen: Gen[Double] = CountGen(stepFunc(currentNumber), stepFunc)
+    override def get: Double = currentNumber
   }
 
-  case class EvoGen(evoFunc: Int => Float, evoCount: Int = 0) extends Gen[Float] {
-    override def nextGen: Gen[Float] = EvoGen(evoFunc, evoCount + 1)
-    override def get: Float = evoFunc(evoCount)
+  case class EvoGen(evoFunc: Int => Double, evoCount: Int = 0) extends Gen[Double] {
+    override def nextGen: Gen[Double] = EvoGen(evoFunc, evoCount + 1)
+    override def get: Double = evoFunc(evoCount)
   }
 
 }
